cloud/requirements.txt
flask==1.1.2
requests==2.24.0
cassandra-driver==3.25.0
